add models for library and cycle modules
