+++ TeSonic/app/models/__init__.py
@@ -4,3 +4,5 @@ from .. import db, login_manager
 from datetime import datetime
 
 from model_auth import *
+from model_library import *
+from model_cycle import *
